cleanup: move value objects to models package
